[Feat] 추천받은 영상 검토 로직 생성

+++ src/main/java/com/kdt/yts/YouSumback/model/entity/UserActivityLog.java
@@ -9,7 +9,7 @@ public class UserActivityLog {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "log_id")
-    private Long logId;
+    private Integer logId;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
